# perf: remove unnecessary `ToList` call and convert linq to code