- Remove unnecessary `ToList` call - 0.15 MB
- Extract a linq call and convert it to code - 0.8 MB


### Benchmarks (timing is inaccurate)
#### Before
| Method                | Mean     | Error   | StdDev   | Gen0      | Gen1      | Allocated |
|---------------------- |---------:|--------:|---------:|----------:|----------:|----------:|
| Default_CodeFormatter | 191.4 ms | 4.05 ms | 11.81 ms | 4000.0000 | 2000.0000 |  42.24 MB |

#### After
| Method                | Mean     | Error    | StdDev   | Median   | Gen0      | Gen1      | Allocated |
|---------------------- |---------:|---------:|---------:|---------:|----------:|----------:|----------:|
| Default_CodeFormatter | 157.3 ms | 11.59 ms | 34.18 ms | 138.8 ms | 4000.0000 | 2000.0000 |  41.29 MB |